refactor: delete posts' index & new and rename form's filename
