+++ apps/posts/templates/posts/show.html
@@ -39,8 +39,8 @@
     </div>
 </div>
 
-<p>Author: {{ post.author }}</p>
 <p>{{ post.content }}</p>
+<small>{{ post.created_at|date:"Y/m/d H:i" }}</small>
 <a href="{% url 'posts:edit' post.id %}">修改評論</a>
 
 <h3>留言</h3>
