+++ apps/posts/urls.py
@@ -5,8 +5,6 @@ from . import views
 app_name = "posts"
 
 urlpatterns = [
-    path("", views.index, name="index"),
-    path("new", views.new, name="new"),
     path("<int:id>", views.show, name="show"),
     path("<int:id>/edit", views.edit, name="edit"),
     path("<int:id>/delete", views.delete, name="delete"),
